- Changed MVP structure. Fragment should be the view, not activity. Fixed
